+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @folionote/config
 
+## 0.1.4
+
+### Patch Changes
+
+- add user number field and update avatar upload limits
+
 ## 0.1.3
 
 ## 0.1.2